Project refactor

+++ Projecteuler/Euler.Logic/EulerHelper.cs
@@ -2,7 +2,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using Euler.Interfaces;
+using Euler.Logic.Interfaces;
 
 namespace Euler
 {
@@ -88,7 +88,7 @@ namespace Euler
         /// </summary>
         /// <param name="maxNumber">Highest number to check</param>
         /// <returns></returns>
-        public IEnumerable<long> ESievePrimeNumbers(int maxNumber)
+        public IEnumerable<long> ESievePrimeNumbers(long maxNumber)
         {
 
             int sieveBound = (int)(maxNumber - 1) / 2;
@@ -96,14 +96,14 @@ namespace Euler
 
             BitArray PrimeBits = new BitArray(sieveBound + 1, true);
 
-            /// Implementing Eratosthenes https://en.wikipedia.org/wiki/Sieve_of_Eratosthenes 
+            // Implementing Eratosthenes https://en.wikipedia.org/wiki/Sieve_of_Eratosthenes 
             for (int i = 1; i <= upperSqrt; i++)
             {
                 if (PrimeBits.Get(i))
                 {
-                    /// Using the Eratosthenes pattern to skip and eliminate numbers as needed.
-                    /// Also skipping the even numbers since they will generally not be prime
-                    /// (The j += 2 * i + 1 ensures we are incrementing by an odd number each time)
+                    // Using the Eratosthenes pattern to skip and eliminate numbers as needed.
+                    // Also skipping the even numbers since they will generally not be prime
+                    // (The j += 2 * i + 1 ensures we are incrementing by an odd number each time)
                     for (int j = i * 2 * (i + 1); j <= sieveBound; j += 2 * i + 1)
                     {
                         PrimeBits.Set(j, false);
@@ -111,14 +111,14 @@ namespace Euler
                 }
             }
 
-            /// Google helped me figure out this part https://en.wikipedia.org/wiki/Logarithm the Math.Log confused me
-            /// Here we are converting our bitarray into a list of integers that we can actually use
+            // Google helped me figure out this part https://en.wikipedia.org/wiki/Logarithm the Math.Log confused me
+            // Here we are converting our bitarray into a list of integers that we can actually use
             List<Int64> numbers = new List<Int64>((int)(maxNumber / (Math.Log(maxNumber) - 1.08366)));
 
-            /// Add "2" because we skipped it earlier
+            // Add "2" because we skipped it earlier
             numbers.Add(2);
 
-            /// Getting the values from our formula and adding to our list
+            // Getting the values from our formula and adding to our list
             for (int i = 1; i <= sieveBound; i++)
             {
                 if (PrimeBits.Get(i))
